mongo description #10

+++ docker/service/accident-recorder-service/src/main/java/ru/smirnov/accidentrecorder/entity/domain/PotentialAccident.java
@@ -39,7 +39,7 @@ public class PotentialAccident {
     private OffsetDateTime accidentDateTime;
 
     @Column(name = "supposed_accuracy", columnDefinition = "NUMERIC(20, 18)", nullable = false)
-    private BigDecimal supposedAccuracy;
+    private Double supposedAccuracy;
 
     @Column(name = "record_reference", columnDefinition = "TEXT", nullable = true)
     private String recordReference;
